Create `default` module for default implementations of Svart traits, test `PartialEq` and `PartialOrd` of Contig implementations
