+++ src/default/mod.rs
@@ -1,0 +1,3 @@
+mod contig;
+
+pub use contig::ContigDefault;
