+++ src/err.rs
@@ -1,4 +1,3 @@
-use std::any::TypeId;
 use std::error::Error;
 use std::fmt::{Display, Formatter};
