+++ src/lib.rs
@@ -2,6 +2,7 @@ mod bound;
 mod coordinate_system;
 mod contig;
 mod err;
+mod default;
 
 // TODO - think about API structure
 pub use self::bound::*;
